refactor(anthropic): split client_test.go and extract assertion helpers
Extract assertPrepareRequestError/assertPrepareRequestSuccess helpers
(mirroring assertNilDepRequired pattern from f8430bce), reducing
TestPrepareRequest_ToAnthropicMessagesError complexity from 12 to 1.

Split the ~1250-line monolithic test file into four focused files:
- client_test.go (457 lines): core send/error/timeout/auth/options
- client_messages_test.go (330 lines): thinking, system, messages
- client_vertex_test.go (367 lines): Vertex AI, caching, metrics
- client_tools_test.go (252 lines): tool schema conversion

Closes #326
